+++ shoppi/src/components/admin/admin.css
@@ -1,0 +1,271 @@
+
+
+.admin-nav{
+    background-color: rgb(231, 231, 231);
+    /* background: linear-gradient(180deg, #DCDCDC 0%, rgba(220, 220, 220, 0) 100%); */
+    width: 100%;
+    height: 8vh;
+    padding-top: 10px;
+}
+
+
+.admin-nav h1{
+    font-size: 2rem;
+    margin-left: 3%;
+    font-weight: 600;
+    display: inline;
+}
+
+.admin-nav button{
+    float: right;
+    margin-right: 3%;
+    background: none;
+    border: none;
+    outline: none;
+    transform: scale(1.3);
+  
+}
+
+.admin-nav button:hover{
+   color: #ffa005;
+
+}
+
+
+.admin-detail{
+    float: left;
+}
+.admin-buttons{
+    float: right;
+    display: inline-flex;
+    margin-top: 25px;
+    margin-right: 2%;
+}
+
+.admin-btn{
+    
+    border: none;
+    outline: none;
+    border-radius: 8px;
+    margin-right: 10px;
+    padding: 5px;
+   
+}
+
+
+.admin-btn:hover{
+    background-color: #ffa005;
+}
+
+
+
+
+
+/* right */
+
+
+.admin-right{
+    /* background-color: #ffa005; */
+    /* height: 92vh; */
+    margin-top: 100px;
+}
+
+.admin-filterBox{
+  /* background-color: white; */
+  width: 85%;
+  margin: auto;
+  margin-top: 30px;
+}
+
+
+.admin-filterBox h2{
+    font-size: 2rem;
+    text-align: center;
+    padding: 2%;
+    color: #ffffff;
+    font-family: inherit;
+    background-color: black;
+    margin: 0;
+}
+
+
+
+.sort-category{
+    /* background-color: rgb(0, 64, 119); */
+    /* background-color: hsla(216, 100%, 59%, 0.493); */
+    /* background-color: hsl(216deg 75% 26% / 13%); */
+    background-color: #ffa005;
+    padding: 3%;
+    font-size: 20px;
+    margin-bottom: 10%;
+}
+
+.sort-category input[type="radio"] {
+    margin: 8px;
+
+  }
+
+.admin-view{
+
+    background-color: #ffffff;
+    color: rgb(82, 82, 82);
+    outline: none;
+    border: none;
+    font-size: 19px;
+    margin-left: 33%;
+    width: 7rem;
+    margin-top: 10%;
+    margin-bottom: 2%;
+    /* padding: 3px 15px 3px 15px; */
+    border-radius: 8px;
+}
+
+
+.admin-view:hover{
+    background-color: #ffa005;
+}
+
+
+
+.admin-order{
+    /* background: #00e747; */
+    padding: 10px;
+    margin-bottom: 10%;
+}
+
+
+.admin-add-product {
+    /* background-color: blueviolet; */
+    margin-bottom: 10%;
+    align-items: center;
+}
+
+
+.addProd-btn{
+    border-radius: 8px;
+    border: none;
+    outline: none;
+    padding: 2%;
+    text-align: center;
+    width: 9rem;
+    background: #65ff05;
+    margin-left: 27%;
+    text-decoration: none;
+    color: black;
+}
+
+.addProd-btn:hover{
+    background-color: #ffa005;
+}
+
+.admin-order{
+    height: 35vh;
+    overflow-y: scroll;
+}
+.order-box{
+    /* background-color: hsla(216, 100%, 59%, 0.493); */
+    background-color: hsl(216deg 75% 26% / 13%);
+    border-radius: 16px;
+    padding: 1%;
+    padding-left: 6%;
+    margin-top: 2%;
+}
+
+
+
+
+/*////////////// add product css ////////////// */
+
+.admin{
+    background-size: 100%;
+    background-image: url("./addIMG.png");
+    /* height: 100vh; */
+    padding-top: 2%;
+}
+
+
+.admin-add{
+
+    /* background-color: #ffa005; */
+    background-color: #3f3f3f47;
+     padding: 2%;
+     border-radius: 16px;
+     margin-left: 34vw;
+     margin-right: 34vw;
+}
+
+.admin-add h2{
+    font-size: 20px;
+}
+
+
+.admin-add-inp{
+    border: none;
+    background-color: #eeeeee;
+    border-radius: 2px;
+    padding-left: 2%;
+    width: 100%;
+    height: 2rem;
+    margin-bottom: 20px;
+    outline: none;
+    box-shadow: inset 4px 4px 4px rgba(0, 0, 0, 0.4);
+} 
+
+.add-btn{
+    background-color: rgb(255, 255, 255);
+    border: none;
+    outline: none;
+    text-align: center;
+    align-items: center;
+    font-size: 25px;
+    padding: 1px 35px 1px 35px;
+    margin-left: 35%;
+    border-radius: 8px;
+}
+
+
+.add-btn:hover{
+    background-color: #65ff05;
+}
+.admin-add button{
+    align-items: center;
+}
+
+
+@media (max-width: 1250px){
+
+
+    .admin-add{
+         margin-left: 30vw;
+         margin-right: 30vw;
+    }
+
+
+
+}
+
+
+@media (max-width: 950px){
+
+
+    .admin-add{
+         margin-left: 24vw;
+         margin-right: 24vw;
+    }
+
+
+
+}
+
+
+@media (max-width: 750px){
+
+
+    .admin-add{
+         margin-left: 20vw;
+         margin-right: 20vw;
+    }
+
+
+
+}
